Crash fixed between Movie Managers
